Combine our two PDFServices into PdfGenerator
